// src/app/funcionario/models/funcionario-filtro.ts
export interface FuncionarioFiltro {
  primeiroNome: string;
  sobrenome: string;
  titulo: string;
  dataNascimento: Date;
  dataAdmissao: Date;
  endereco: string;
  cidade: string;
  estado: string;
  pais: string;
  cep: string;
  fone: string;
  fax: string;
  email: string;
  gerenteId: number;
}